integration tests improvements
