+++ kit/dapp/test/system/access-manager/revoke-role.spec.ts
@@ -1,5 +1,5 @@
 // @vitest-environment node
-import { ORPCError } from "@orpc/server";
+import type { AccessControlRoles } from "@/lib/fragments/the-graph/access-control-fragment";
 import { getAddress } from "viem";
 import { beforeAll, describe, expect, it } from "vitest";
 import { getOrpcClient } from "../../utils/orpc-client";
@@ -10,14 +10,14 @@ import {
   signInWithUser,
 } from "../../utils/user";
 
-describe("Access Manager - Grant Role ORPC routes", () => {
+describe("Access Manager - Revoke Role ORPC routes", () => {
   let adminClient: ReturnType<typeof getOrpcClient>;
   let investorClient: ReturnType<typeof getOrpcClient>;
 
   const testAddresses = {
-    valid1: getAddress("0x1234567890123456789012345678901234567890"),
-    valid2: getAddress("0x2345678901234567890123456789012345678901"),
-    valid3: getAddress("0x3456789012345678901234567890123456789012"),
+    valid1: getAddress("0xabcdefabcdefabcdefabcdefabcdefabcdefabcd"),
+    valid2: getAddress("0xbcdefabcdefabcdefabcdefabcdefabcdefabcde"),
+    valid3: getAddress("0xcdefabcdefabcdefabcdefabcdefabcdefabcdef"),
     invalid: "0xinvalid",
   };
 
@@ -27,11 +27,33 @@ describe("Access Manager - Grant Role ORPC routes", () => {
 
     const investorHeaders = await signInWithUser(DEFAULT_INVESTOR);
     investorClient = getOrpcClient(investorHeaders);
-  });
 
-  describe("successful role grants", () => {
-    it("should grant a single role to a single account", async () => {
-      const result = await adminClient.system.grantRole({
+    // Grant roles to the test addresses
+    const rolesToGrant: AccessControlRoles[] = [
+      "systemManager",
+      "complianceManager",
+      "tokenManager",
+    ];
+    for (const role of rolesToGrant) {
+      console.log(`Granting ${role} to test addresses`);
+      await adminClient.system.grantRole({
+        verification: {
+          verificationCode: DEFAULT_PINCODE,
+          verificationType: "pincode",
+        },
+        accounts: [
+          testAddresses.valid1,
+          testAddresses.valid2,
+          testAddresses.valid3,
+        ],
+        role,
+      });
+    }
+  }, 60_000);
+
+  describe("successful role revokes", () => {
+    it("should revoke a single role from a single account", async () => {
+      const result = await adminClient.system.revokeRole({
         verification: {
           verificationCode: DEFAULT_PINCODE,
           verificationType: "pincode",
@@ -51,11 +73,11 @@ describe("Access Manager - Grant Role ORPC routes", () => {
         (role) => role.account === testAddresses.valid1
       );
       expect(updatedSystemRoles).toBeDefined();
-      expect(updatedSystemRoles?.roles).toContain("tokenManager");
+      expect(updatedSystemRoles?.roles).not.toContain("tokenManager");
     });
 
-    it("should grant a role to multiple accounts", async () => {
-      const result = await adminClient.system.grantRole({
+    it("should revoke a role from multiple accounts", async () => {
+      const result = await adminClient.system.revokeRole({
         verification: {
           verificationCode: DEFAULT_PINCODE,
           verificationType: "pincode",
@@ -80,17 +102,14 @@ describe("Access Manager - Grant Role ORPC routes", () => {
         excludeContracts: true,
       });
       const updatedSystemRoles = systemRoles.find(
-        (role) =>
-          role.account === testAddresses.valid1 ||
-          role.account === testAddresses.valid2 ||
-          role.account === testAddresses.valid3
+        (role) => role.account === testAddresses.valid1
       );
       expect(updatedSystemRoles).toBeDefined();
-      expect(updatedSystemRoles?.roles).toContain("complianceManager");
+      expect(updatedSystemRoles?.roles).not.toContain("complianceManager");
     });
 
     it("should handle empty accounts array", async () => {
-      const result = await adminClient.system.grantRole({
+      const result = await adminClient.system.revokeRole({
         verification: {
           verificationCode: DEFAULT_PINCODE,
           verificationType: "pincode",
@@ -106,8 +125,8 @@ describe("Access Manager - Grant Role ORPC routes", () => {
   });
 
   describe("permission validation", () => {
-    it("should allow admin users to grant roles", async () => {
-      const result = await adminClient.system.grantRole({
+    it("should allow admin users to revoke roles", async () => {
+      const result = await adminClient.system.revokeRole({
         verification: {
           verificationCode: DEFAULT_PINCODE,
           verificationType: "pincode",
@@ -121,9 +140,9 @@ describe("Access Manager - Grant Role ORPC routes", () => {
       });
     });
 
-    it("should reject non-admin users from granting roles", async () => {
+    it("should reject non-admin users from revoking roles", async () => {
       await expect(
-        investorClient.system.grantRole({
+        investorClient.system.revokeRole({
           verification: {
             verificationCode: DEFAULT_PINCODE,
             verificationType: "pincode",
@@ -132,10 +151,7 @@ describe("Access Manager - Grant Role ORPC routes", () => {
           role: "tokenManager",
         })
       ).rejects.toThrow(
-        new ORPCError("USER_NOT_AUTHORIZED", {
-          message:
-            "User does not have the required role to execute this action.",
-        })
+        "User does not have the required role to execute this action."
       );
     });
   });
@@ -143,7 +159,7 @@ describe("Access Manager - Grant Role ORPC routes", () => {
   describe("error handling", () => {
     it("should reject invalid role names", async () => {
       await expect(
-        adminClient.system.grantRole({
+        adminClient.system.revokeRole({
           verification: {
             verificationCode: DEFAULT_PINCODE,
             verificationType: "pincode",
@@ -151,12 +167,12 @@ describe("Access Manager - Grant Role ORPC routes", () => {
           accounts: [testAddresses.valid1],
           role: "invalidRole" as never,
         })
-      ).rejects.toThrow(new ORPCError("INPUT_VALIDATION_FAILED"));
+      ).rejects.toThrow("INPUT_VALIDATION_FAILED");
     });
 
     it("should reject invalid wallet addresses", async () => {
       await expect(
-        adminClient.system.grantRole({
+        adminClient.system.revokeRole({
           verification: {
             verificationCode: DEFAULT_PINCODE,
             verificationType: "pincode",
@@ -164,12 +180,12 @@ describe("Access Manager - Grant Role ORPC routes", () => {
           accounts: [testAddresses.invalid],
           role: "tokenManager",
         })
-      ).rejects.toThrow(new ORPCError("INPUT_VALIDATION_FAILED"));
+      ).rejects.toThrow("INPUT_VALIDATION_FAILED");
     });
 
     it("should reject mixed valid and invalid addresses", async () => {
       await expect(
-        adminClient.system.grantRole({
+        adminClient.system.revokeRole({
           verification: {
             verificationCode: DEFAULT_PINCODE,
             verificationType: "pincode",
@@ -181,12 +197,12 @@ describe("Access Manager - Grant Role ORPC routes", () => {
           ],
           role: "tokenManager",
         })
-      ).rejects.toThrow(new ORPCError("INPUT_VALIDATION_FAILED"));
+      ).rejects.toThrow("INPUT_VALIDATION_FAILED");
     });
 
     it("should reject incorrect pincode verification", async () => {
       await expect(
-        adminClient.system.grantRole({
+        adminClient.system.revokeRole({
           verification: {
             verificationCode: "000000",
             verificationType: "pincode",
@@ -194,17 +210,13 @@ describe("Access Manager - Grant Role ORPC routes", () => {
           accounts: [testAddresses.valid1],
           role: "tokenManager",
         })
-      ).rejects.toThrow(
-        new ORPCError("PORTAL_ERROR", {
-          message: "Invalid authentication challenge",
-        })
-      );
+      ).rejects.toThrow("Invalid authentication challenge");
     });
   });
 
   describe("edge cases", () => {
     it("should handle duplicate accounts in the array", async () => {
-      const result = await adminClient.system.grantRole({
+      const result = await adminClient.system.revokeRole({
         verification: {
           verificationCode: DEFAULT_PINCODE,
           verificationType: "pincode",
