+++ kit/dapp/test/token/token-create.spec.ts
@@ -1,4 +1,3 @@
-import { ORPCError } from "@orpc/server";
 import { from } from "dnum";
 import { describe, expect, test } from "vitest";
 import { getOrpcClient } from "../utils/orpc-client";
@@ -70,9 +69,7 @@ describe("Token create", () => {
         countryCode: "056", // Belgium numeric code for testing
       })
     ).rejects.toThrow(
-      new ORPCError("USER_NOT_AUTHORIZED", {
-        message: "User does not have the required role to execute this action.",
-      })
+      "User does not have the required role to execute this action."
     );
   });
 });
